improved performance of SymbolCard by adding components and memoizing them
