+++ frontend/package.json
@@ -22,7 +22,7 @@
   },
   "devDependencies": {
     "@types/node": "^22.5.4",
-    "@types/react": "^18.0.28",
+    "@types/react": "latest",
     "@types/react-dom": "^18.0.11",
     "@types/vite-plugin-react-svg": "^0.2.2",
     "@vitejs/plugin-react": "^3.1.0",
