+++ frontend/src/components/ListItem/ListItem.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { memo } from 'react';
 import './listItem.css';
 type ListItemProps = {
   Icon: React.ReactNode;
@@ -23,4 +23,4 @@ const ListItem = ({ Icon, label, spacing }: ListItemProps) => {
   );
 };
 
-export default ListItem;
+export default memo(ListItem);
