+++ frontend/src/components/SymbolCard/SymbolCardBody/SymbolCardBody.tsx
@@ -1,42 +1,17 @@
-import ListItem from '@/components/ListItem';
 import './symbolCardBody.css';
-import { ReactComponent as CompanyIcon } from '@/assets/company.svg';
-import { ReactComponent as IndustryIcon } from '@/assets/industry.svg';
-import { ReactComponent as MarketCapIcon } from '@/assets/market_cap.svg';
 import SymbolCardPrice from '@/components/SymbolCard/SymbolCardPrice';
 import { Stock } from '@/store/stocksSlice';
-import { showShortenedAmount } from '@/helpers/currency';
+import SymbolCardList from '@/components/SymbolCard/SymbolCardList';
 
-type SymbolCardBodyProps = {
+export type SymbolCardBodyProps = {
   price: number;
 } & Pick<Stock, 'companyName' | 'industry' | 'marketCap'>;
 
-const SymbolCardBody = ({ price, companyName, industry, marketCap }: SymbolCardBodyProps) => {
-  const listItems = [
-    {
-      id: 'companyName',
-      label: companyName,
-      icon: CompanyIcon
-    },
-    {
-      id: 'industry',
-      label: industry,
-      icon: IndustryIcon
-    },
-    {
-      id: 'marketCap',
-      label: showShortenedAmount(marketCap),
-      icon: MarketCapIcon
-    }
-  ];
-
+const SymbolCardBody = ({ price, ...rest }: SymbolCardBodyProps) => {
   return (
     <div className="symbolCard__body">
       <SymbolCardPrice price={price} />
-      {listItems.map(({ id, label, icon }) => {
-        const Icon = icon;
-        return <ListItem key={id} Icon={<Icon />} label={label} spacing="space-between" />;
-      })}
+      <SymbolCardList {...rest} />
     </div>
   );
 };
