+++ frontend/src/components/SymbolCard/SymbolCardList/index.tsx
@@ -1,0 +1,3 @@
+import SymbolCardList from './SymbolCardList';
+
+export default SymbolCardList;
